Fit spots to a new Visium-like array

Given transformed pixel coordinates, modify the 'array_row' and 'array_col' columns to represent a larger Visium capture area containing all capture areas in a common coordinate system. The number of array rows/cols generally changes from the Visium standards of 78 and 128 (and even may change in ratio between num rows and num cols).

.fit_to_array(coords, inter_spot_dist_px)

Arguments

coords

A tibble() whose rows represent capture areas of the same group, and containing columns 'array_row', 'array_col', 'pxl_row_in_fullres', and 'pxl_col_in_fullres'.

inter_spot_dist_px

numeric(1) vector giving the pixel distance between any 2 spots in the new coordinates.

Value

A tibble() with modified 'array_row' + 'array_col' columns, as well as new 'pxl_row_in_fullres_rounded' and 'pxl_col_in_fullres_rounded' columns representing the pixel coordinates rounded to the nearest exact array coordinates.

Details

Runtime is O(n) with the number of spots, making it much faster than say, a distance-matrix-based approach running at O(n^2).

Author

Nicholas J. Eagles